backend/pyproject.toml
[project]
name = "ai-startup-validator-backend"
version = "0.1.0"
description = "AI Startup Idea Validator - Multi-Agent Backend"
requires-python = ">=3.10"
dependencies = [
    "genkit>=0.9.0",
    "genkit-google-genai>=0.9.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.9.0",
    "google-generativeai>=0.8.0",
    "firebase-admin>=6.5.0",
    "google-cloud-dlp>=3.18.0",
    "sse-starlette>=2.1.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.backends.legacy:build"
